Create block abstraction and generic block
