--- plugins/woocommerce/src/Internal/Admin/BlockTemplates/BlockContainer.php
@@ -1,49 +1,0 @@
-<?php
-
-namespace Automattic\WooCommerce\Internal\Admin\BlockTemplates;
-
-use Automattic\WooCommerce\Admin\BlockTemplates\BlockContainerInterface;
-use Automattic\WooCommerce\Admin\BlockTemplates\BlockInterface;
-use Automattic\WooCommerce\Admin\BlockTemplates\BlockTemplateInterface;
-
-/**
- * Block configuration used to specify blocks in BlockTemplate.
- */
-class BlockContainer extends Block implements BlockContainerInterface {
-	use BlockContainerTrait;
-
-	/**
-	 * Get the block configuration as a formatted template.
-	 *
-	 * @return array The block configuration as a formatted template.
-	 */
-	public function get_formatted_template(): array {
-		$arr = [
-			$this->get_name(),
-			$this->get_attributes(),
-		];
-
-		$inner_blocks = $this->get_inner_blocks_sorted_by_order();
-
-		if ( ! empty( $inner_blocks ) ) {
-			$arr[] = array_map(
-				function( BlockInterface $block ) {
-					return $block->get_formatted_template();
-				},
-				$inner_blocks
-			);
-		}
-
-		return $arr;
-	}
-
-	/**
-	 * Generate a block ID based on a base.
-	 *
-	 * @param string $id_base The base to use when generating an ID.
-	 * @return string
-	 */
-	public function &add_block( array $block_config ): BlockInterface {
-		return $this->add_inner_block( $block_config );
-	}
-}
